+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Changed
 - Upgrade Gradle to 6.8.2
+- Dependencies:
+  - Upgrade `android-junit5` to `1.7.1.1`
 
 ## [1.1.0-RC1]
 ### Changed